Extract inference multiprocessing context
