+++ docs/REFACTOR_AUDIT.md
@@ -61,6 +61,7 @@ Date: 2026-05-07
 - Extracted order-preserving streaming drain helpers into `fiberhmm/inference/streaming_drain.py`; streaming apply and fused apply+recall now share drain behavior outside `parallel.py`.
 - Extracted streaming worker initializers and chunk worker entry points into `fiberhmm/inference/streaming_workers.py`, while keeping compatibility imports in `parallel.py`.
 - Extracted region-parallel worker initializers and BAM/BED/fused region worker entry points into `fiberhmm/inference/region_workers.py`, while keeping compatibility imports in `parallel.py`.
+- Extracted multiprocessing start-method selection into `fiberhmm/inference/mp_context.py`, while keeping compatibility imports in `parallel.py`.
 - Closed inline posterior writers from apply processing `finally` blocks in both streaming and legacy paths, with failure-path regression coverage.
 - Closed fused DAF streaming reference FASTA handles from the processing `finally` block, including failure-path coverage.
 - Extracted region-parallel posterior TSV formatting, output-path resolution, and merge ordering into `fiberhmm/posteriors/region_tsv.py` with direct coverage.
@@ -269,12 +270,19 @@ Date: 2026-05-07
 - `python -m compileall -q fiberhmm tests`: passed.
 - `python -m pytest`: 370 passed, 26 deselected in 12.29s.
 - `python -m pytest -m benchmark tests/benchmarks`: 26 passed in 51.76s.
+- `python -m ruff check fiberhmm/inference/parallel.py fiberhmm/inference/mp_context.py tests/test_inference_parallel.py`: passed.
+- `python -m compileall -q fiberhmm/inference/parallel.py fiberhmm/inference/mp_context.py tests/test_inference_parallel.py`: passed.
+- `python -m pytest tests/test_inference_parallel.py tests/test_region_cleanup.py tests/test_streaming_pipeline.py tests/test_call_pipeline.py tests/test_mode_equivalence.py`: 78 passed in 14.10s.
+- `python -m ruff check fiberhmm tests`: passed.
+- `python -m compileall -q fiberhmm tests`: passed.
+- `python -m pytest`: 371 passed, 26 deselected in 11.91s.
+- `python -m pytest -m benchmark tests/benchmarks`: 26 passed in 53.07s.
 
 ## Current Shape
 
 Largest tracked Python files:
 
-- `fiberhmm/inference/parallel.py`: 1472 lines.
+- `fiberhmm/inference/parallel.py`: 1436 lines.
 - `fiberhmm/cli/extract_tags.py`: 1395 lines.
 - `fiberhmm/core/bam_reader.py`: 1340 lines.
 - `fiberhmm/core/hmm.py`: 1123 lines.
@@ -292,7 +300,7 @@ Ignored local build artifacts exist (`build/`, `fiberhmm.egg-info/`) but are not
 
 2. Continue shrinking `fiberhmm/inference/parallel.py`.
 
-   The first shared tagging/unification slice, streaming read-filter slice, active BAM output helper extractions, region worker contracts, aggregation helpers, higher-level worker cleanup tests, fused stage-boundary extraction, streaming drain extraction, streaming worker extraction, and region worker extraction are complete, but the module still mixes process lifecycle, posterior export, and orchestration. The next work should shift toward measured speed/stability improvements while continuing to keep `parallel.py` shrinking.
+   The first shared tagging/unification slice, streaming read-filter slice, active BAM output helper extractions, region worker contracts, aggregation helpers, higher-level worker cleanup tests, fused stage-boundary extraction, streaming drain extraction, streaming worker extraction, region worker extraction, and multiprocessing context extraction are complete, but the module still mixes process lifecycle, posterior export, and orchestration. The next work should shift toward measured speed/stability improvements while continuing to keep `parallel.py` shrinking.
 
 3. Add remaining `fiberhmm-call` characterization tests.
 
@@ -365,6 +373,7 @@ Phase 3: pipeline split.
 - Split `parallel.py` into streaming, region, workers, and orchestration modules while preserving `process_bam_for_footprints` as the public compatibility wrapper.
 - Extract streaming worker initialization and chunk worker entry points; done for the streaming apply and fused apply+recall worker functions.
 - Extract region worker initialization and per-region worker entry points; done for apply BAM, BED, and fused BAM workers.
+- Extract shared multiprocessing context selection; done.
 - Keep old import paths re-exporting during the transition.
 
 Phase 4: performance work.
